Style adjustments in U2U (referrals shadow + label links)
